Separate yosupo from segtree2
