+++ crates/algolib/segtree2/Cargo.toml
@@ -10,7 +10,6 @@ edition = "2018"
 alg_traits = { path = "../../algebra/alg_traits" }
 rand = "0.7.3"
 prettydiff = "0.3.1"
-yansi = "0.5.0"
 
 [dev-dependencies.query_test]
 git = "https://github.com/ngtkana/query_test.git"
@@ -20,7 +19,7 @@ package = "query_test"
 [dev-dependencies]
 fp = { path = "../../utils/fp" }
 test_vector2 = { path = "../../test_tools/test_vector2" }
+yosupo = { path = "../../test_tools/yosupo" }
 alg_inversion_number = { path = "../../algebra/alg_inversion_number" }
 ngtio = { path = "../../utils/ngtio" }
 assert-impl = "0.1.3"
-prettydiff = "0.3.1"
