+++ crates/test_tools/yosupo/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "yosupo"
+version = "0.1.0"
+authors = ["ngtkana <ngtkana@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+yansi = "0.5.0"
+prettydiff = "0.3.1"
